Applying comments from review
