+++ packages/shopify-app-remix/src/server/.eslintrc.js
@@ -7,7 +7,7 @@ module.exports = {
           {
             group: ['../react/*'],
             message:
-              'Importing from src/react in src/server is not allowed as they must be kept separate',
+              'Importing from src/react in src/server is not allowed as they must be kept separate, because it breaks Remix bundling.',
           },
         ],
       },
