+++ packages/shopify-app-remix/src/server/__tests__/test-helper.ts
@@ -1,23 +1,46 @@
 import '@shopify/shopify-api/adapters/web-api';
-import '../../adapters/node';
+import '../adapters/node';
 import crypto from 'crypto';
 
 import jwt from 'jsonwebtoken';
-import {JwtPayload, Session} from '@shopify/shopify-api';
-import {SessionStorage} from '@shopify/shopify-app-session-storage';
-
-import {APP_BRIDGE_URL} from '../auth/const';
 import {
-  API_KEY,
-  API_SECRET_KEY,
-  APP_URL,
-  BASE64_HOST,
-  TEST_SHOP,
-  testConfig,
-} from '../../__tests__/test-helper';
+  LATEST_API_VERSION,
+  JwtPayload,
+  LogSeverity,
+  Session,
+} from '@shopify/shopify-api';
+import {SessionStorage} from '@shopify/shopify-app-session-storage';
+import {MemorySessionStorage} from '@shopify/shopify-app-session-storage-memory';
 
-export * from '../../__tests__/test-helper';
+import {AppConfigArg} from '../config-types';
+import {APP_BRIDGE_URL} from '../auth/const';
 
+export function testConfig(
+  overrides: Partial<AppConfigArg> = {},
+): AppConfigArg & {sessionStorage: SessionStorage} {
+  return {
+    apiKey: API_KEY,
+    apiSecretKey: API_SECRET_KEY,
+    scopes: ['testScope'],
+    apiVersion: LATEST_API_VERSION,
+    appUrl: APP_URL,
+    logger: {
+      log: jest.fn(),
+      level: LogSeverity.Debug,
+    },
+    isEmbeddedApp: true,
+    sessionStorage: new MemorySessionStorage(),
+    ...overrides,
+  };
+}
+
+export const API_SECRET_KEY = 'testApiSecretKey';
+export const API_KEY = 'testApiKey';
+export const APP_URL = 'https://my-test-app.myshopify.io';
+export const SHOPIFY_HOST = 'totally-real-host.myshopify.io';
+export const BASE64_HOST = Buffer.from(SHOPIFY_HOST).toString('base64');
+export const TEST_SHOP = 'test-shop.myshopify.com';
+export const GRAPHQL_URL = `https://${TEST_SHOP}/admin/api/${LATEST_API_VERSION}/graphql.json`;
 const USER_ID = 12345;
 
 interface TestJwt {
