+++ packages/shopify-app-remix/src/server/adapters/node/__tests__/node-app-bridge-url.test.ts
@@ -1,5 +1,5 @@
-import {appBridgeUrl} from '../../../server/auth/helpers/app-bridge-url';
-import {APP_BRIDGE_URL} from '../../../server/auth/const';
+import {appBridgeUrl} from '../../../auth/helpers/app-bridge-url';
+import {APP_BRIDGE_URL} from '../../../auth/const';
 
 describe('node setup import', () => {
   /* eslint-disable no-process-env */
