+++ packages/shopify-app-remix/src/server/adapters/node/index.ts
@@ -5,7 +5,7 @@ import {
   setCrypto,
 } from '@shopify/shopify-api/runtime';
 
-import {setAppBridgeUrlOverride} from '../../server/auth/helpers';
+import {setAppBridgeUrlOverride} from '../../auth/helpers';
 
 setCrypto(crypto as any);
